[design] 🏀 Nav바 색 변화 및 애니메이션 추가
