+++ app/(head)/page.tsx
@@ -4,6 +4,7 @@ import ProgressBar from '@/components/molecules/GoalProgress';
 import { LowestProductList } from '@/components/molecules/LowestProductList';
 import { AdCarousel } from '@/components/molecules/sion/AdCarousel';
 import { AccentText } from '@/components/ui/AccentText';
+import { MOCK_LOWEST_PRODUCT } from '@/constants/mockdata';
 import { Triangle } from '@/public/icons/Triangle';
 import { ArcElement, Chart as ChartJS, Legend, Tooltip } from 'chart.js';
 
@@ -15,16 +16,6 @@ export type Goal = {
   saved_amount: number;
   start_date: string;
   end_date: string;
-};
-
-type LowestProduct = {
-  productId: number;
-  productName: string;
-  productImage: string;
-  price: number;
-  storeName: string;
-  link: string;
-  discountRate: number;
 };
 
 type UserData = {
@@ -48,53 +39,6 @@ export default function Home() {
     saved_amount: 18320,
     spent_amount: -31710,
   };
-
-  const MOCK_LOWEST_PRODUCT: LowestProduct[] = [
-    {
-      productId: 12345,
-      productName:
-        '나드 리프레쉬 퍼퓸드 바디워시 본품, 프레쉬라벤더향, 680ml, 1개입',
-      productImage:
-        'https://img.danawa.com/prod_img/500000/426/012/img/2012426_1.jpg?shrink=330:*&_v=2024082713372',
-      price: 8800,
-      storeName: '11번가',
-      link: 'https://www.11st.co.kr/products/7588965357?service_id=estimatedn&utm_term=&utm_campaign=%B4%D9%B3%AA%BF%CDpc_%B0%A1%B0%DD%BA%F1%B1%B3%B1%E2%BA%BB&utm_source=%B4%D9%B3%AA%BF%CD_PC_PCS&utm_medium=%B0%A1%B0%DD%BA%F1%B1%B3',
-      discountRate: 11.0,
-    },
-    {
-      productId: 12345,
-      productName:
-        '나드 리프레쉬 퍼퓸드 바디워시 본품, 프레쉬라벤더향, 680ml, 1개입',
-      productImage:
-        'https://img.danawa.com/prod_img/500000/426/012/img/2012426_1.jpg?shrink=330:*&_v=20240827133727',
-      price: 8800,
-      storeName: '11번가',
-      link: 'https://www.11st.co.kr/products/7588965357?service_id=estimatedn&utm_term=&utm_campaign=%B4%D9%B3%AA%BF%CDpc_%B0%A1%B0%DD%BA%F1%B1%B3%B1%E2%BA%BB&utm_source=%B4%D9%B3%AA%BF%CD_PC_PCS&utm_medium=%B0%A1%B0%DD%BA%F1%B1%B3',
-      discountRate: 11.0,
-    },
-    {
-      productId: 12345,
-      productName:
-        '나드 리프레쉬 퍼퓸드 바디워시 본품, 프레쉬라벤더향, 680ml, 1개입',
-      productImage:
-        'https://img.danawa.com/prod_img/500000/426/012/img/2012426_1.jpg?shrink=330:*&_v=2024082713372',
-      price: 8800,
-      storeName: '11번가',
-      link: 'https://www.11st.co.kr/products/7588965357?service_id=estimatedn&utm_term=&utm_campaign=%B4%D9%B3%AA%BF%CDpc_%B0%A1%B0%DD%BA%F1%B1%B3%B1%E2%BA%BB&utm_source=%B4%D9%B3%AA%BF%CD_PC_PCS&utm_medium=%B0%A1%B0%DD%BA%F1%B1%B3',
-      discountRate: 11.0,
-    },
-    {
-      productId: 12345,
-      productName:
-        '나드 리프레쉬 퍼퓸드 바디워시 본품, 프레쉬라벤더향, 680ml, 1개입',
-      productImage:
-        'https://img.danawa.com/prod_img/500000/426/012/img/2012426_1.jpg?shrink=330:*&_v=20240827133727',
-      price: 8800,
-      storeName: '11번가',
-      link: 'https://www.11st.co.kr/products/7588965357?service_id=estimatedn&utm_term=&utm_campaign=%B4%D9%B3%AA%BF%CDpc_%B0%A1%B0%DD%BA%F1%B1%B3%B1%E2%BA%BB&utm_source=%B4%D9%B3%AA%BF%CD_PC_PCS&utm_medium=%B0%A1%B0%DD%BA%F1%B1%B3',
-      discountRate: 11.0,
-    },
-  ];
 
   return (
     <div className='flex justify-center flex-col items-center w-full'>
