+++ components/molecules/LowestProductList.tsx
@@ -17,9 +17,9 @@ type LowestProductListProps = {
 export const LowestProductList = ({ products }: LowestProductListProps) => {
   return (
     <div className='flex space-x-4 overflow-x-scroll'>
-      {products.map((product) => (
+      {products.map((product, index) => (
         <LowestProductCard
-          key={product.productId}
+          key={index}
           imageUrl={product.productImage}
           title={product.productName}
           discount={product.discountRate}
