+++ constants/mockdata.ts
@@ -33,6 +33,16 @@ export type PaymentData = {
   account_name: string;
 };
 
+type LowestProduct = {
+  productId: number;
+  productName: string;
+  productImage: string;
+  price: number;
+  storeName: string;
+  link: string;
+  discountRate: number;
+};
+
 export const MOCK_STAT: Stat = {
   date: '2025-01',
   total_spent: 832000,
@@ -103,3 +113,50 @@ export const MOCK_PAYMENT_DETAIL = {
     },
   ],
 };
+
+export const MOCK_LOWEST_PRODUCT: LowestProduct[] = [
+  {
+    productId: 12345,
+    productName:
+      '나드 리프레쉬 퍼퓸드 바디워시 본품, 프레쉬라벤더향, 680ml, 1개입',
+    productImage:
+      'https://img.danawa.com/prod_img/500000/426/012/img/2012426_1.jpg?shrink=330:*&_v=2024082713372',
+    price: 8800,
+    storeName: '11번가',
+    link: 'https://www.11st.co.kr/products/7588965357?service_id=estimatedn&utm_term=&utm_campaign=%B4%D9%B3%AA%BF%CDpc_%B0%A1%B0%DD%BA%F1%B1%B3%B1%E2%BA%BB&utm_source=%B4%D9%B3%AA%BF%CD_PC_PCS&utm_medium=%B0%A1%B0%DD%BA%F1%B1%B3',
+    discountRate: 11.0,
+  },
+  {
+    productId: 12345,
+    productName:
+      '나드 리프레쉬 퍼퓸드 바디워시 본품, 프레쉬라벤더향, 680ml, 1개입',
+    productImage:
+      'https://img.danawa.com/prod_img/500000/426/012/img/2012426_1.jpg?shrink=330:*&_v=20240827133727',
+    price: 8800,
+    storeName: '11번가',
+    link: 'https://www.11st.co.kr/products/7588965357?service_id=estimatedn&utm_term=&utm_campaign=%B4%D9%B3%AA%BF%CDpc_%B0%A1%B0%DD%BA%F1%B1%B3%B1%E2%BA%BB&utm_source=%B4%D9%B3%AA%BF%CD_PC_PCS&utm_medium=%B0%A1%B0%DD%BA%F1%B1%B3',
+    discountRate: 11.0,
+  },
+  {
+    productId: 12345,
+    productName:
+      '나드 리프레쉬 퍼퓸드 바디워시 본품, 프레쉬라벤더향, 680ml, 1개입',
+    productImage:
+      'https://img.danawa.com/prod_img/500000/426/012/img/2012426_1.jpg?shrink=330:*&_v=2024082713372',
+    price: 8800,
+    storeName: '11번가',
+    link: 'https://www.11st.co.kr/products/7588965357?service_id=estimatedn&utm_term=&utm_campaign=%B4%D9%B3%AA%BF%CDpc_%B0%A1%B0%DD%BA%F1%B1%B3%B1%E2%BA%BB&utm_source=%B4%D9%B3%AA%BF%CD_PC_PCS&utm_medium=%B0%A1%B0%DD%BA%F1%B1%B3',
+    discountRate: 11.0,
+  },
+  {
+    productId: 12345,
+    productName:
+      '나드 리프레쉬 퍼퓸드 바디워시 본품, 프레쉬라벤더향, 680ml, 1개입',
+    productImage:
+      'https://img.danawa.com/prod_img/500000/426/012/img/2012426_1.jpg?shrink=330:*&_v=20240827133727',
+    price: 8800,
+    storeName: '11번가',
+    link: 'https://www.11st.co.kr/products/7588965357?service_id=estimatedn&utm_term=&utm_campaign=%B4%D9%B3%AA%BF%CDpc_%B0%A1%B0%DD%BA%F1%B1%B3%B1%E2%BA%BB&utm_source=%B4%D9%B3%AA%BF%CD_PC_PCS&utm_medium=%B0%A1%B0%DD%BA%F1%B1%B3',
+    discountRate: 11.0,
+  },
+];
